Restructure tests and add lockfree baseline/perf hardening
