+++ minimalstdlib/test/correctness/array_test.cpp
