+++ minimalstdlib/test/correctness/elimination_array_test.cpp
