+++ minimalstdlib/test/correctness/fixed_string_test.cpp
