+++ minimalstdlib/test/correctness/lockfree_bitblock_set_test.cpp
@@ -397,83 +397,6 @@ namespace
         CHECK_EQUAL(MT_SIZE, blocks.total_bits_set());
     }
 
-    namespace
-    {
-        minstd::atomic<bool> perf_start{false};
-
-        struct alignas(64) perf_thread_args
-        {
-            minstd::bitblock_set<16384> *blocks_ = nullptr;
-            uint64_t rng_state_ = 0;
-            size_t iterations_ = 0;
-        };
-
-        void *perf_worker(void *arg)
-        {
-            auto *args = static_cast<perf_thread_args *>(arg);
-
-            while (!perf_start.load(minstd::memory_order_acquire))
-            {
-            }
-
-            for (size_t i = 0; i < args->iterations_; ++i)
-            {
-                auto result = args->blocks_->acquire_first_available();
-
-                if (result.has_value())
-                {
-                    args->blocks_->release(result.value());
-                }
-            }
-
-            return nullptr;
-        }
-    }
-
-    TEST(BitblockSetTests, AcquireFirstAvailableThreadScaling)
-    {
-        constexpr size_t MAX_THREADS = 32;
-        constexpr size_t ITERATIONS_PER_THREAD = 2000000;
-
-        minstd::bitblock_set<16384> blocks;
-
-        pthread_t threads[MAX_THREADS];
-        perf_thread_args arguments[MAX_THREADS];
-
-        printf("Bitblock_set acquire_first_available/release ops/sec:\n");
-
-        for (size_t num_threads = 1; num_threads <= MAX_THREADS; ++num_threads)
-        {
-            perf_start.store(false, minstd::memory_order_release);
-
-            for (size_t i = 0; i < num_threads; ++i)
-            {
-                arguments[i].blocks_ = &blocks;
-                arguments[i].rng_state_ = 0x9e3779b97f4a7c15ULL ^ (i * 0x100000001b3ULL);
-                arguments[i].iterations_ = ITERATIONS_PER_THREAD;
-
-                pthread_create(&threads[i], nullptr, perf_worker, &arguments[i]);
-            }
-
-            timespec start_time{};
-            timespec end_time{};
-            clock_gettime(CLOCK_MONOTONIC, &start_time);
-            perf_start.store(true, minstd::memory_order_release);
-
-            for (size_t i = 0; i < num_threads; ++i)
-            {
-                pthread_join(threads[i], nullptr);
-            }
-
-            clock_gettime(CLOCK_MONOTONIC, &end_time);
-            double duration = static_cast<double>(end_time.tv_sec - start_time.tv_sec);
-            duration += static_cast<double>(end_time.tv_nsec - start_time.tv_nsec) / 1e9;
-            double ops_per_sec = static_cast<double>(ITERATIONS_PER_THREAD * num_threads) / duration;
-
-            printf("  %2zu threads: %f\n", num_threads, ops_per_sec);
-        }
-    }
-
     TEST(BitblockSetTests, OrphanBitBugMultiWordAcquireFirstWordConflict)
     {
         // This test demonstrates the orphan bit bug discovered by TLA+ analysis.
