+++ minimalstdlib/test/performance/composite_resource_performance_test.cpp
