+++ minimalstdlib/test/performance/cpputest_performance_main.cpp
@@ -1,0 +1,8 @@
+// Copyright 2024 Stephan Friedl. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+#include <CppUTest/TestHarness.h>
+#include <CppUTest/CommandLineTestRunner.h>
+
+CPPUTEST_DEFAULT_MAIN
