+++ minimalstdlib/test/shared/strnstr.cpp
